Added StatusCode to ApiException message
When an API call fails, the resulting message provides very little information if there is no content (as was the case with a recent error from the API). This made it impossible to tell the general context of the error, i.e. is it a true API error, is it rate limiting, is it a 400 bad request, etc. without specifically handling an ApiException in code and looking at those properties. Since we're already putting details in the exception message, we should also consider putting the HttpStatus in there as well as that is a first-class citizen of a response.